+++ Freyja/Module/HomeController.swift
@@ -41,3 +41,4 @@ extension HomeController {
         ].forEach({ $0.isActive = true })
     }
 }
+
